Update signup.html
PHP code to allow the user to register.  It checks if the user hit the register button and goes through the correct error checks before submitting their information.  It's also for checking if a user hits the login button after typing in their credentials(Not tested), We don't have the html form for login yet.